Removed storage dependency from eventhubs
