+++ sdk/eventhubs/CMakeLists.txt
@@ -9,3 +9,4 @@ set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_WINDOWS_EXPORT_ALL_SYMBOLS ON)
 
 add_subdirectory(azure-messaging-eventhubs)
+add_subdirectory(azure-messaging-eventhubs-checkpointstore-blob)
