+++ sdk/eventhubs/azure-messaging-eventhubs-checkpointstore-blob/test/eventhubs_test_base.hpp
@@ -1,0 +1,21 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+#include "gtest/gtest.h"
+
+#include <azure/core/test/test_base.hpp>
+
+class EventHubsTestBase : public Azure::Core::Test::TestBase {
+public:
+  EventHubsTestBase() { TestBase::SetUpTestSuiteLocal(AZURE_TEST_ASSETS_DIR); }
+  // Create
+  virtual void SetUp() override
+  {
+    Azure::Core::Test::TestBase::SetUpTestBase(AZURE_TEST_RECORDING_DIR);
+  }
+  virtual void TearDown() override
+  {
+    // Make sure you call the base classes TearDown method to ensure recordings are made.
+    TestBase::TearDown();
+  }
+};
