+++ sdk/eventhubs/azure-messaging-eventhubs-checkpointstore-blob/vcpkg/Config.cmake.in
@@ -1,0 +1,13 @@
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT License.
+
+@PACKAGE_INIT@
+
+include(CMakeFindDependencyMacro)
+find_dependency(azure-core-amqp-cpp)
+find_dependency(azure-messaging-eventhubs-cpp)
+find_dependency(azure-storage-blobs-cpp)
+
+include("${CMAKE_CURRENT_LIST_DIR}/azure-messaging-eventhubs-checkpointstore-blob-cppTargets.cmake")
+
+check_required_components("azure-messaging-eventhubs-checkpointstore-blob-cpp")
