+++ sdk/eventhubs/azure-messaging-eventhubs-checkpointstore-blob/vcpkg/vcpkg.json
@@ -1,0 +1,38 @@
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT License.
+
+{
+  "name": "azure-messaging-eventhubs-checkpointstore-blob-cpp",
+  "version-semver": "@AZ_LIBRARY_VERSION@",
+  "description": [
+    "Microsoft Azure Messaging Event Hubs SDK for C++ Blob Checkpoint Store",
+    "This library provides an Azure-Storage-Blobs based implementation of an Azure Messaging Event Hubs SDK Checkpoint Store."
+  ],
+  "homepage": "https://github.com/Azure/azure-sdk-for-cpp/tree/main/sdk/eventhubs/azure-messaging-eventhubs",
+  "license": "MIT",
+  "dependencies": [
+    {
+      "name": "azure-core-amqp-cpp",
+      "default-features": false,
+      "version>=": "1.0.0-beta.2"
+    },
+    {
+      "name": "azure-messaging-eventhubs",
+      "default-features": false,
+      "version>=": "1.0.0-beta.2"
+    },
+    {
+      "name": "azure-storage-blobs-cpp",
+      "default-features": false,
+      "version>=": "12.8.0"
+    },
+    {
+      "name": "vcpkg-cmake",
+      "host": true
+    },
+    {
+      "name": "vcpkg-cmake-config",
+      "host": true
+    }
+  ]
+}
